backend/requirements.txt
pydantic==2.12.5
fastapi==0.128.0
ezdxf==1.4.3
uvicorn==0.40.0
matplotlib==3.10.0
torch>=2.2.0
transformers>=4.44.0
accelerate>=0.30.0
sentencepiece>=0.2.0
pytest==9.0.2
bcrypt>=4.0.1,<5
google-auth>=2.38.0,<3
python-multipart>=0.0.9,<1
